Add exception handling for User

+++ src/main/java/se/iths/corkdork/controller/UserController.java
@@ -1,9 +1,12 @@
 package se.iths.corkdork.controller;
 
+import org.jetbrains.annotations.NotNull;
 import se.iths.corkdork.entity.UserEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import se.iths.corkdork.exception.BadRequestException;
+import se.iths.corkdork.exception.EntityNotFoundException;
 import se.iths.corkdork.service.UserService;
 
 import java.util.Optional;
@@ -18,12 +21,14 @@ public class UserController {
         this.userService = userService;
     }
 
-    @PostMapping("signup")
+    @PostMapping()
     public ResponseEntity<UserEntity> createUser(@RequestBody UserEntity userEntity){
+        if(userEntity.getFirstName().isEmpty() || userEntity.getLastName().isEmpty() || userEntity.getUserName().isEmpty()
+        || userEntity.getPassword().isEmpty() || userEntity.getEmail().isEmpty())
+            throw new BadRequestException("Every user credential is mandatory");
+
         UserEntity createdUser = userService.createUser(userEntity);
-
         return new ResponseEntity<>(createdUser, HttpStatus.CREATED);
-
     }
 
     @PutMapping
@@ -37,13 +42,18 @@ public class UserController {
 
     @DeleteMapping("{id}")
     public ResponseEntity<Void> deleteUser(@PathVariable Long id){
+        if(userService.findUserById(id).isEmpty())
+            throw new EntityNotFoundException(notFound(id));
+
         userService.deleteUser(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-
     }
 
     @GetMapping("{id}")
     public ResponseEntity<Optional<UserEntity>> findUserById(@PathVariable Long id){
+        if(userService.findUserById(id).isEmpty())
+            throw new EntityNotFoundException(notFound(id));
+
         Optional<UserEntity> foundUser = userService.findUserById(id);
         return new ResponseEntity<>(foundUser, HttpStatus.OK);
     }
@@ -51,8 +61,15 @@ public class UserController {
     @GetMapping
     public ResponseEntity<Iterable<UserEntity>> findAllUsers(){
         Iterable<UserEntity> allUsers = userService.findAllUsers();
+        if(!allUsers.iterator().hasNext())
+            throw new EntityNotFoundException("Failed to find any wines.");
+
         return new ResponseEntity<>(allUsers, HttpStatus.OK);
     }
 
+    @NotNull
+    private String notFound(Long id) {
+        return "User with ID: " + id + " was not found.";
+    }
 
 }
